requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyYAML==5.4.1
Werkzeug==1.0.1
